requirements.txt
streamlit
psycopg2-binary
pandas
google-auth
google-api-python-client